move unit tests into their own files
