+++ tests/unit/models/ledger_objects/test_amendments.py
@@ -1,0 +1,47 @@
+from unittest import TestCase
+
+from xrpl.models.ledger_objects.amendments import Amendments, Majority
+from xrpl.models.ledger_objects.ledger_object import LedgerObject
+
+
+class TestAmendments(TestCase):
+    def test_amendments(self):
+        amendment_json = {
+            "Majorities": [
+                {
+                    "Majority": {
+                        "Amendment": "1562511F573A19AE9BD103B5D6B9E01B3B46805AEC"
+                        "5D3C4805C902B51"
+                        "4399146",
+                        "CloseTime": 535589001,
+                    }
+                }
+            ],
+            "Amendments": [
+                "42426C4D4F1009EE67080A9B7965B44656D7714D104A72F9B4369F97ABF044EE",
+                "4C97EBA926031A7CF7D7B36FDE3ED66DDA5421192D63DE53FFB46E43B9DC8373",
+                "6781F8368C4771B83E8B821D88F580202BCB4228075297B19E4FDC5233F1EFDC",
+                "740352F2412A9909880C23A559FCECEDA3BE2126FED62FC7660D628A06927F11",
+            ],
+            "Flags": 0,
+            "LedgerEntryType": "Amendments",
+            "index": "7DB0788C020F02780A673DC74757F23823FA3014C1866E72CC4CD8B226CD6EF4",
+        }
+        actual = LedgerObject.from_xrpl(amendment_json)
+        expected = Amendments(
+            index="7DB0788C020F02780A673DC74757F23823FA3014C1866E72CC4CD8B226CD6EF4",
+            amendments=[
+                "42426C4D4F1009EE67080A9B7965B44656D7714D104A72F9B4369F97ABF044EE",
+                "4C97EBA926031A7CF7D7B36FDE3ED66DDA5421192D63DE53FFB46E43B9DC8373",
+                "6781F8368C4771B83E8B821D88F580202BCB4228075297B19E4FDC5233F1EFDC",
+                "740352F2412A9909880C23A559FCECEDA3BE2126FED62FC7660D628A06927F11",
+            ],
+            majorities=[
+                Majority(
+                    amendment="1562511F573A19AE9BD103B5D6B9E01B3B46805AEC5D3C"
+                    "4805C902B514399146",
+                    close_time=535589001,
+                )
+            ],
+        )
+        self.assertEqual(actual, expected)
